Garbage Collector: Move gc related functions to `gc.c`
