+++ common.h
@@ -14,11 +14,13 @@
 #define DEBUG_PRINT_CODE
 #define DEBUG_LOG_GC
 #define EXCESSIVE_GC_MODE
+#define RUN_GC_AT_END
 
 #undef DEBUG_TRACE_EXECUTION
 #undef DEBUG_PRINT_CODE
-#undef DEBUG_LOG_GC
 #undef EXCESSIVE_GC_MODE
+#undef RUN_GC_AT_END
+#undef DEBUG_LOG_GC
 
 
 #endif
