+++ vm/gc.h
@@ -19,4 +19,18 @@ typedef struct{
 
 void addObject(Object*);
 
+void markObjects();
+void markRoots();
+void markCompilerRoots();
+
+void sweepObjects();
+
+void markValue(Value value);
+void markObject(Object* object);
+void markHashTable(Table* table);
+void markStack();
+void markCallFrame();
+
+void addChildObjectsToGCQueue(Object*);
+void freeStringsFromVMHashTable();
 #endif
